requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pwdlib==0.3.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
PyJWT==2.12.1
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
